[refactor] rename tokens.rb to access_tokens.rb
