--- lib/tokens.rb
@@ -1,24 +1,0 @@
-module Tokens
-
-  class << self
-
-    def digest(string)
-      if ActiveModel::SecurePassword.min_cost
-        cost = BCrypt::Engine::MIN_COST
-      else
-        cost = BCrypt::Engine.cost
-      end
-      BCrypt::Password.create(string, cost: cost)
-    end
-
-    def new_token
-      SecureRandom.urlsafe_base64
-    end
-
-    def new_access_token(user)
-      "#{user.id}:#{new_token}"
-    end
-
-  end
-
-end
